Switch to using an accessor rather than a mandatory field to make timelocks optional
