+++ crates/utils/sov-test-utils/Cargo.toml
@@ -51,7 +51,6 @@ sov-modules-rollup-blueprint = { workspace = true, features = ["native"] }
 sov-bank = { workspace = true, features = ["native", "test-utils"] }
 sov-accounts = { workspace = true, features = ["native"] }
 sov-uniqueness = { workspace = true, features = ["native"] }
-sov-timelock = { workspace = true, features = ["native"] }
 sov-kernels = { workspace = true, features = ["native"] }
 sov-ledger-apis = { workspace = true }
 sov-api-spec = { workspace = true }
@@ -83,7 +82,6 @@ arbitrary = [
 	"sov-address/arbitrary",
 	"sov-accounts/arbitrary",
 	"sov-uniqueness/arbitrary",
-	"sov-timelock/arbitrary",
 	"sov-mock-da/arbitrary",
 	"sov-bank/arbitrary",
 	"sov-paymaster/arbitrary",
